Fixed Intel Driver Bug and downgrade to OpenGL 3.3
Intel OpenGL Drivers are bad at guessing, so i defined the attributes properly now and everything should work

+++ src/main/java/io/github/paul1365972/simulation/world/InteractionHandler.java
@@ -2,5 +2,5 @@ package io.github.paul1365972.simulation.world;
 
 @FunctionalInterface
 public interface InteractionHandler {
-	void interact(Particle t, Particle o, float deltaT);
+	void interact(Particle t, Particle o, double deltaT);
 }
